+++ test/external/external_test_opt.py
@@ -4,7 +4,6 @@ import numpy as np
 import torch
 
 from tinygrad import nn, GlobalCounters, Tensor, Device
-from tinygrad.helpers import getenv
 from tinygrad.nn import optim
 from tinygrad.nn.state import get_parameters
 from tinygrad.engine.realize import capturing
@@ -31,11 +30,6 @@ class CLCache:
     if self.allowed is not None:
       assert self.count == self.allowed, f"{self.count} != {self.allowed}"
 
-from extra.models.convnext import ConvNeXt
-from extra.models.efficientnet import EfficientNet
-from extra.models.resnet import ResNet18
-from extra.models.vit import ViT
-
 @unittest.skipUnless(Device.DEFAULT == "GPU", "Not Implemented")
 class TestInferenceMinKernels(unittest.TestCase):
   def setUp(self):
@@ -43,45 +37,6 @@ class TestInferenceMinKernels(unittest.TestCase):
     Tensor.training = False
   def tearDown(self):
     Tensor.training = self.training_old
-
-  @unittest.skipIf(not PUSH_PERMUTES, "this test requires PUSH_PERMUTES")
-  def test_convnext(self):
-    model = ConvNeXt()
-    for p in get_parameters(model): p.assign(np.zeros(p.shape, dtype=p.dtype.np))
-    img = Tensor.randn(1, 3, 224, 224)
-    with CLCache(129):
-      model(img).realize()
-
-  def test_enet(self):
-    model = EfficientNet(getenv("ENET_NUM", 0), has_se=False)
-    for p in get_parameters(model): p.assign(np.zeros(p.shape, dtype=p.dtype.np))
-    img = Tensor.randn(1, 3, 224, 224)
-    with CLCache(51):
-      model.forward(img).realize()
-
-  def test_enet_se(self):
-    model = EfficientNet(getenv("ENET_NUM", 0), has_se=True)
-    for p in get_parameters(model): p.assign(np.zeros(p.shape, dtype=p.dtype.np))
-    img = Tensor.randn(1, 3, 224, 224)
-    # TODO: this seems very high
-    with CLCache(115):
-      model.forward(img).realize()
-
-  def test_resnet(self):
-    model = ResNet18()
-    for p in get_parameters(model): p.assign(np.zeros(p.shape, dtype=p.dtype.np))
-    img = Tensor.randn(1, 3, 224, 224)
-    with CLCache(23):
-      model.forward(img).realize()
-
-  def test_vit(self):
-    model = ViT(embed_dim=192, num_heads=3)
-    for p in get_parameters(model): p.assign(np.zeros(p.shape, dtype=p.dtype.np))
-    img = Tensor.randn(1, 3, 224, 224)
-    with CLCache(209) as cache: # NOTE: this is way too high
-      out = model.forward(img)
-      assert cache.count == 0, "ViT prerealized?"
-      out.realize()
 
   @unittest.skip("llama is fp16 but CI does not have fp16")
   def test_llama(self):
@@ -182,49 +137,6 @@ class TestOpt(unittest.TestCase):
         img_bn = bn(img).realize()
         print(img_bn)
         assert cache.count == 3, f"optimizer didn't fold batchnorm, got {cache.count}"
-
-  def test_fold_conv_sgd(self):
-    with Tensor.train():
-      img = Tensor.ones(2,3,4,4)
-      c1 = nn.Conv2d(3,32,3)
-      opt = optim.SGD(get_parameters(c1))
-      with CLCache() as cache:
-        opt.zero_grad()
-        c1(img).relu().sum().backward()
-        opt.step()
-        assert cache.count == 5, f"optimizer didn't fold conv-backward SGD, got {cache.count}"
-
-  def test_fold_conv_adam(self):
-    with Tensor.train():
-      img = Tensor.ones(2,3,4,4)
-      c1 = nn.Conv2d(3,32,3)
-      opt = optim.Adam(get_parameters(c1), lr=1e-4)
-      with CLCache(allowed=10):
-        opt.zero_grad()
-        c1(img).relu().sum().backward()
-        opt.step()
-
-  def test_fold_2convs_adam(self):
-    with Tensor.train():
-      img = Tensor.ones(2,3,64,64)
-      c1 = nn.Conv2d(3,16,3,bias=False)
-      c2 = nn.Conv2d(16,32,3,bias=False)
-      opt = optim.Adam(get_parameters([c1, c2]), lr=1e-4)
-      with CLCache(allowed=13):
-        opt.zero_grad()
-        c2(c1(img).relu()).relu().sum().backward()
-        opt.step()
-
-  def test_fold_2convs_sgd(self):
-    with Tensor.train():
-      img = Tensor.ones(2,3,64,64)
-      c1 = nn.Conv2d(3,16,3,bias=False)
-      c2 = nn.Conv2d(16,32,3,bias=False)
-      opt = optim.SGD(get_parameters([c1, c2]))
-      with CLCache(allowed=8):
-        opt.zero_grad()
-        c2(c1(img).relu()).relu().sum().backward()
-        opt.step()
 
   def test_fold_2convs_sgd_nesterov_momentum_wd(self):
     with Tensor.train():
